Chat with PDFs/requirements.txt
langchain
PyPDF2
python-dotenv
streamlit==1.4.0
faiss-cpu==1.7.4
altair==4
tiktoken
huggingface-hub
InstructorEmbedding
sentence-transformers
